refactoring stuff in preparation to add google support
